change activator signture/contract
Change activators such that they take an outputStream argument,
that might be null. If non-null, that stream object should be
used for output. If null, activator should create its own output
stream. Either way, output stream is still returned.

This is going to be necessary in the near future to allow
on-become syntax to do our version of "tail call optimization".

Also, started differentiating stream tempos.